fix structure of HomePage Components
Destructuring the file Components/HomePage into separate component files: DailyNorma, ImageOfBottle, and WaterRatioPanel.
