+++ src/components/HomePage/WaterRatioPanel.jsx
@@ -1,0 +1,55 @@
+import React, { useState } from 'react';
+import iconSprite from '../../images/SVG/symbol-defs.svg';
+import {
+  RangeAndAddWater,
+  RangeDiv,
+  RangeTitle,
+  StyledRangeInput,
+  PercentageDiv,
+  PercentageOfRange,
+  BoldPercentageOfRange,
+  ButtonAddWater,
+  SvgButton,
+} from './HomePage.styled';
+export const WaterRatioPanel = () => {
+  const [sliderValue, setSliderValue] = useState(0);
+
+  const handleChange = event => {
+    const value = event.target.value;
+    setSliderValue(value);
+  };
+
+  const progress = (sliderValue / 100) * 100;
+
+  const sliderStyle = {
+    background: `linear-gradient(to right, #9EBBFF ${progress}%, #D7E3FF ${progress}%)`,
+  };
+  return (
+    <RangeAndAddWater>
+      <RangeDiv>
+        <RangeTitle>Today</RangeTitle>
+        <StyledRangeInput
+          type="range"
+          min="0"
+          max="100"
+          value={sliderValue}
+          onChange={handleChange}
+          style={sliderStyle}
+        ></StyledRangeInput>
+        <PercentageDiv>
+          <PercentageOfRange>0%</PercentageOfRange>
+          <PercentageOfRange>
+            <BoldPercentageOfRange>50%</BoldPercentageOfRange>
+          </PercentageOfRange>
+          <PercentageOfRange>100%</PercentageOfRange>
+        </PercentageDiv>
+      </RangeDiv>
+      <ButtonAddWater>
+        <SvgButton>
+          <use href={iconSprite + '#icon-plus-circle'} />
+        </SvgButton>
+        Add Water
+      </ButtonAddWater>
+    </RangeAndAddWater>
+  );
+};
